requirements.txt
Adafruit-Blinka==8.46.1
adafruit-circuitpython-dht==4.0.4
azure-servicebus==7.12.2
python-dotenv==1.0.1
